Let the HTTP gateway read data directly from the supervisor.
This eliminates the need to write JSON dat files out to disk, and read
those in every time the gateway needs to serve a request. Instead, the
gateway and the supervisor share a common data structure, which the
supervisor updates periodically. The gateway just reads from that
directly when it needs to serve requests. This is not only more
performant, but it also eliminates the security risk of having secrets
storing unencrypted on disk.

Signed-off-by: Josh Black <raskchanky@gmail.com>